+++ src/components/layout/Header.tsx
@@ -1,9 +1,19 @@
 import { useEffect, useState } from 'react';
-import * as S from '../../styledComponent/styledHeader/StHeader';
-import BasicHeader from './BasicHeader';
-import ChangeHeader from './ChangeHeader';
+import * as S from '../../styledComponent/styledLayout/StHeader';
+import { useNavigate } from 'react-router-dom';
 
 const Header = () => {
+  const navigate = useNavigate();
+
+  const homeClickHandler = () => {
+    navigate('/');
+  };
+  const loginHandler = () => {
+    navigate('/');
+  };
+  const wishListHandler = () => {
+    navigate('/');
+  };
   const [scrollPostion, setScrollPoistion] = useState<number>(0);
   const updateScroll = () => {
     setScrollPoistion(window.scrollY || document.documentElement.scrollTop);
@@ -13,8 +23,19 @@ const Header = () => {
   });
 
   return (
-    <S.HeaderContainer>
-      {scrollPostion < 150 ? <BasicHeader /> : <ChangeHeader />};
+    <S.HeaderContainer color={scrollPostion < 150 ? 'transparent' : 'black'}>
+      <S.Wrapper>
+        <S.Header>
+          <S.BrandLogo onClick={homeClickHandler}>FancyPlace</S.BrandLogo>
+        </S.Header>
+        <S.Header>
+          <S.HeaderButton>
+            <button onClick={loginHandler}>로그인</button>
+            <button onClick={wishListHandler}>장바구니</button>
+            <button>KoR \</button>
+          </S.HeaderButton>
+        </S.Header>
+      </S.Wrapper>
     </S.HeaderContainer>
   );
 };
